Priyaranjan billing system

+++ src/main/java/com/app/billingsystem/repository/ItemRepository.java
@@ -1,7 +1,10 @@
 package com.app.billingsystem.repository;
 
 import com.app.billingsystem.models.entities.Item;
+import jakarta.validation.constraints.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item,Integer> {
+import java.util.Optional;
+
+public interface ItemRepository extends JpaRepository<Item, Long> {
 }
